+++ src/app/configuration/route.ts
@@ -1,5 +1,6 @@
-import { configuration } from "@/lib/configuration";
+import { buildConfiguration } from "./route-configuration";
 
-export async function GET(_: Request) {
-  return Response.json(configuration);
+export async function GET() {
+  const DOMAIN = process.env.APP_URL ?? process.env.VIRKAILIJA_URL ?? 'https://localhost:3404';
+  return Response.json(buildConfiguration(DOMAIN));
 }
