+++ src/components/providers/configuration-provider.tsx
@@ -2,13 +2,16 @@
 import { FullSpinner } from '@/components/full-spinner';
 import { ErrorView } from '../error-view';
 import { useConfiguration } from '@/hooks/useConfiguration';
+import { createContext } from 'react';
+
+export const ConfigurationContext = createContext<Record<string, any>>({configuration: {}});
 
 export function ConfigurationProvider({
   children,
 }: {
   children: React.ReactNode;
 }) {
-  const { isLoading, isError, error, refetch } = useConfiguration();
+  const { isLoading, isError, error, refetch, data } = useConfiguration();
 
   switch (true) {
     case isLoading:
@@ -16,6 +19,6 @@ export function ConfigurationProvider({
     case isError:
       return <ErrorView error={error} reset={refetch} />;
     default:
-      return <>{children}</>;
+      return <ConfigurationContext.Provider value={{configuration: data}}>{children}</ConfigurationContext.Provider>;
   }
 }
