+++ src/lib/ataru/ataru-service.ts
@@ -4,11 +4,11 @@ import {
   Hakukelpoisuus,
   Maksuvelvollisuus,
 } from './ataru-types';
-import { configuration } from '../configuration';
 import { client } from '../http-client';
 import { Language } from '../localization/localization-types';
 import { queryOptions } from '@tanstack/react-query';
 import { KoutaOidParams } from '../kouta/kouta-types';
+import { getConfiguration } from '@/hooks/useConfiguration';
 
 const getMaksuvelvollisuus = (toive?: {
   hakukohdeOid: string;
@@ -92,6 +92,7 @@ export async function getAtaruHakemukset({
   henkiloOid,
   henkilotunnus,
 }: GetHakijatParams) {
+  const configuration = await getConfiguration();
   const url = new URL(configuration.hakemuksetUrl);
   if (hakuOid) {
     url.searchParams.append('hakuOid', hakuOid);
