+++ src/lib/localization/tolgee-config.ts
@@ -1,4 +1,4 @@
-import { configuration, isTesting, localTranslations } from '../configuration';
+import { isTesting, localTranslations } from '../configuration';
 import { BackendFetch, DevTools, Tolgee } from '@tolgee/react';
 import { FormatIcu } from '@tolgee/format-icu';
 
@@ -26,6 +26,24 @@ export function TolgeeBase() {
     return tg
       .use(
         BackendFetch({
+          prefix: 'https://localhost:3404/lokalisointi/tolgee',
+          next: {
+            revalidate: REVALIDATE_TIME_SECONDS,
+          },
+        }),
+      )
+      .use(DevTools())
+      .updateDefaults({
+        apiKey,
+        apiUrl,
+        defaultNs: NAMESPACE,
+        ns: [NAMESPACE],
+        projectId: 11100,
+      });
+    /* return getConfiguration().then(configuration => {
+      return tg
+      .use(
+        BackendFetch({
           prefix: configuration.lokalisointiUrl,
           next: {
             revalidate: REVALIDATE_TIME_SECONDS,
@@ -38,5 +56,6 @@ export function TolgeeBase() {
         ns: [NAMESPACE],
         projectId: 11100,
       });
+    }); */
   }
 }
