+++ src/lib/valintalaskentakoostepalvelu/valintalaskentakoostepalvelu-service.ts
@@ -1,4 +1,3 @@
-import { configuration } from '../configuration';
 import {
   abortableClient,
   client,
@@ -59,11 +58,13 @@ import { AssertionError } from 'assert';
 import { queryOptions } from '@tanstack/react-query';
 import { Language } from '../localization/localization-types';
 import { HakemuksenValinnanTulos } from '../valinta-tulos-service/valinta-tulos-types';
+import { getConfiguration } from '@/hooks/useConfiguration';
 
 export const getHakukohteenValintatuloksetIlmanHakijanTilaa = async (
   hakuOid: string,
   hakukohdeOid: string,
 ): Promise<Array<HenkilonValintaTulos>> => {
+  const configuration = await getConfiguration();
   const { data } = await client.get<Array<{ tila: string; hakijaOid: string }>>(
     `${configuration.valintalaskentaKoostePalveluUrl}proxy/valintatulosservice/ilmanhakijantilaa/haku/${hakuOid}/hakukohde/${hakukohdeOid}`,
   );
@@ -92,6 +93,7 @@ export type PisteetForHakemus = {
 };
 
 const getPisteetForHakemus = async ({ hakemusOid }: { hakemusOid: string }) => {
+  const configuration = await getConfiguration();
   const res = await client.get<PisteetForHakemus>(
     configuration.koostetutPistetiedotHakemukselleUrl({ hakemusOid }),
   );
@@ -138,6 +140,7 @@ export const getPisteetForHakukohde = async (
   hakuOid: string,
   hakukohdeOid: string,
 ): Promise<HakukohteenPistetiedot> => {
+  const configuration = await getConfiguration();
   const kokeetPromise = getValintakoeAvaimetHakukohteelle(hakukohdeOid);
   const hakemuksetPromise = getHakemukset({ hakuOid, hakukohdeOid });
   const pisteTiedotFetch = abortableClient.get<{
@@ -195,6 +198,7 @@ export const updatePisteetForHakukohde = async (
   hakukohdeOid: string,
   pistetiedot: Array<HakemuksenPistetiedot>,
 ) => {
+  const configuration = await getConfiguration();
   const mappedPistetiedot = pistetiedot.map((p) => {
     const additionalData = pipe(
       p.valintakokeenPisteet,
@@ -227,6 +231,7 @@ const getValintakoeOsallistumiset = async ({
 }: {
   hakukohdeOid: string;
 }) => {
+  const configuration = await getConfiguration();
   const response = await client.get<Array<HakutoiveValintakoeOsallistumiset>>(
     configuration.valintakoeOsallistumisetUrl({ hakukohdeOid }),
   );
@@ -348,6 +353,7 @@ const pollDocumentProcess = async (
   processId: string,
   infiniteWait: boolean,
 ) => {
+  const configuration = await getConfiguration();
   let pollTimes = 10;
 
   while (pollTimes || infiniteWait) {
@@ -395,6 +401,7 @@ const processDocumentAndReturnDocumentId = async (
 };
 
 export const downloadReadyProcessDocument = async (dokumenttiId: string) => {
+  const configuration = await getConfiguration();
   const documentRes = await client.get<Blob>(
     configuration.lataaDokumenttiUrl({ dokumenttiId }),
   );
@@ -418,6 +425,7 @@ export const getValintakoeExcel = async ({
   hakemusOids,
   valintakoeTunniste,
 }: GetValintakoeExcelParams & { valintakoeTunniste: Array<string> }) => {
+  const configuration = await getConfiguration();
   const urlWithQuery = new URL(configuration.startExportValintakoeExcelUrl);
   urlWithQuery.searchParams.append('hakuOid', hakuOid);
   urlWithQuery.searchParams.append('hakukohdeOid', hakukohdeOid);
@@ -441,6 +449,7 @@ export const getValintakoeOsoitetarrat = async ({
   hakemusOids,
   valintakoeTunniste,
 }: GetValintakoeOsoitetarratParams) => {
+  const configuration = await getConfiguration();
   const urlWithQuery = new URL(
     configuration.startExportValintakoeOsoitetarratUrl,
   );
@@ -465,6 +474,7 @@ export const getOsoitetarratHakemuksille = async ({
   tag: string;
   hakemusOids: Array<string>;
 }) => {
+  const configuration = await getConfiguration();
   const startProcessResponse = await client.post<{ id: string }>(
     configuration.startExportOsoitetarratHakemuksilleUrl,
     {
@@ -481,6 +491,7 @@ export const getOsoitetarratHaulle = async ({
 }: {
   hakuOid: string;
 }) => {
+  const configuration = await getConfiguration();
   const startProcessResponse = await client.post<{ id: string }>(
     `${configuration.startExportOsoitetarratHaulleUrl}?hakuOid=${hakuOid}`,
     '',
@@ -494,6 +505,7 @@ export const getValintalaskennanTulosExcel = async ({
 }: {
   hakukohdeOid: string;
 }) => {
+  const configuration = await getConfiguration();
   const excelRes = await client.get<Blob>(
     configuration.valintalaskennanTulosExcelUrl({ hakukohdeOid }),
   );
@@ -509,6 +521,7 @@ export const getValintatapajonoTulosExcel = async ({
   hakukohdeOid,
   valintatapajonoOid,
 }: ValintatapaJonoTulosExcelProps) => {
+  const configuration = await getConfiguration();
   const urlWithQuery = new URL(
     configuration.startExportValintatapajonoTulosExcelUrl,
   );
@@ -524,6 +537,7 @@ export const getValintatapajonoTulosExcel = async ({
 };
 
 const pollDocumentSeuranta = async (uuid: string) => {
+  const configuration = await getConfiguration();
   let pollTimes = 10;
 
   while (pollTimes) {
@@ -564,6 +578,7 @@ export const saveValintatapajonoTulosExcel = async ({
   valintatapajonoOid,
   file,
 }: ValintatapaJonoTulosExcelProps & { file: File }) => {
+  const configuration = await getConfiguration();
   const urlWithQuery = new URL(
     configuration.startImportValintatapajonoTulosExcelUrl,
   );
@@ -588,6 +603,7 @@ export const getPistesyottoExcel = async ({
   hakuOid,
   hakukohdeOid,
 }: KoutaOidParams) => {
+  const configuration = await getConfiguration();
   const urlWithQuery = new URL(configuration.startExportPistesyottoExcelUrl);
   urlWithQuery.searchParams.append('hakuOid', hakuOid);
   urlWithQuery.searchParams.append('hakukohdeOid', hakukohdeOid);
@@ -607,6 +623,7 @@ export const getSijoittelunTulosExcel = async ({
 }: KoutaOidParams & {
   sijoitteluajoId: string;
 }) => {
+  const configuration = await getConfiguration();
   const urlWithQuery = new URL(configuration.sijoittelunTulosExcelUrl);
   urlWithQuery.searchParams.append('hakuOid', hakuOid);
   urlWithQuery.searchParams.append('hakukohdeOid', hakukohdeOid);
@@ -623,6 +640,7 @@ export const getSijoittelunTulosExcel = async ({
 export const getSijoittelunTulosHaulleExcel = async (
   hakuOid: string,
 ): Promise<FileResult> => {
+  const configuration = await getConfiguration();
   const urlWithQuery = new URL(configuration.sijoittelunTulosHaulleExcelUrl);
   urlWithQuery.searchParams.append('hakuOid', hakuOid);
   const createResponse = await client.post<{ id: string }>(
@@ -641,6 +659,7 @@ export const savePistesyottoExcel = async ({
 }: KoutaOidParams & {
   excelFile: File;
 }) => {
+  const configuration = await getConfiguration();
   const urlWithQuery = new URL(configuration.startImportPistesyottoUrl);
   urlWithQuery.searchParams.append('hakuOid', hakuOid);
   urlWithQuery.searchParams.append('hakukohdeOid', hakukohdeOid);
@@ -677,6 +696,7 @@ export const getHarkinnanvaraisuudetHakemuksille = async ({
 }: {
   hakemusOids: Array<string>;
 }) => {
+  const configuration = await getConfiguration();
   const res = await client.post<Array<HakemuksenHarkinnanvaraisuustiedot>>(
     configuration.harkinnanvaraisuudetHakemuksilleUrl,
     hakemusOids,
@@ -699,6 +719,7 @@ export const luoHyvaksymiskirjeetPDF = async ({
   deadline?: Date | null;
   onlyForbidden: boolean;
 }): Promise<string> => {
+  const configuration = await getConfiguration();
   const hakukohdeNimi = translateName(hakukohde.nimi);
   const opetuskieliCode = (getOpetuskieliCode(hakukohde) || 'fi').toUpperCase();
   const pvm = deadline
@@ -745,6 +766,7 @@ export const luoHyvaksymiskirjeetHaullePDF = async ({
   lang?: Language;
   templateName?: string;
 }): Promise<FileResult> => {
+  const configuration = await getConfiguration();
   const urlWithQuery = new URL(configuration.hyvaksymiskirjeetUrl);
   urlWithQuery.searchParams.append('hakuOid', hakuOid);
   urlWithQuery.searchParams.append('vainTulosEmailinKieltaneet', 'false');
@@ -772,6 +794,7 @@ export const luoEiHyvaksymiskirjeetPDF = async ({
   hakukohde: Hakukohde;
   letterBody: string;
 }): Promise<string> => {
+  const configuration = await getConfiguration();
   const opetuskieliCode = (getOpetuskieliCode(hakukohde) || 'fi').toUpperCase();
   const urlWithQuery = new URL(configuration.eihyvaksymiskirjeetUrl);
   urlWithQuery.searchParams.append('hakuOid', hakukohde.hakuOid);
@@ -802,6 +825,7 @@ export const luoEiHyvaksymiskirjeetPDFHaulle = async ({
   templateName: string;
   lang: Language;
 }): Promise<FileResult> => {
+  const configuration = await getConfiguration();
   const urlWithQuery = new URL(configuration.jalkiohjauskirjeetUrl);
   urlWithQuery.searchParams.append('hakuOid', hakuOid);
   urlWithQuery.searchParams.append('tag', hakuOid);
@@ -826,6 +850,7 @@ export const luoOsoitetarratHakukohteessaHyvaksytyille = async ({
   sijoitteluajoId: string;
   hakukohde: Hakukohde;
 }): Promise<string> => {
+  const configuration = await getConfiguration();
   const urlWithQuery = new URL(
     configuration.startExportOsoitetarratSijoittelussaHyvaksytyilleUrl,
   );
@@ -852,6 +877,7 @@ export const getKirjepohjatHakukohteelle = async (
   kirjepohjanNimi: KirjepohjaNimi,
   hakukohde: Hakukohde,
 ): Promise<Array<Kirjepohja>> => {
+  const configuration = await getConfiguration();
   const opetuskieliCode = (getOpetuskieliCode(hakukohde) || 'fi').toUpperCase();
   const res = await client.get<Array<TemplateResponse>>(
     configuration.kirjepohjat({
@@ -886,6 +912,7 @@ export const getDocumentIdForHakukohde = async (
   hakukohdeOid: string,
   documentType: DokumenttiTyyppi,
 ): Promise<string | null> => {
+  const configuration = await getConfiguration();
   const res = await client.get<[{ documentId: string }] | undefined>(
     configuration.dokumentitUrl({ tyyppi: documentType, hakukohdeOid }),
   );
@@ -903,7 +930,7 @@ export const getMyohastyneetHakemukset = async ({
   if (hakemusOids.length === 0) {
     return [];
   }
-
+  const configuration = await getConfiguration();
   const response = await client.post<
     Array<{ hakemusOid: string; mennyt: boolean; vastaanottoDeadline: string }>
   >(
@@ -920,6 +947,7 @@ export const getMyohastyneetHakemukset = async ({
 export const tuloskirjeidenMuodostuksenTilanne = async (
   hakuOid: string,
 ): Promise<Array<LetterCounts>> => {
+  const configuration = await getConfiguration();
   const res = await client.get<{
     [key: string]: {
       [key: string]: {
@@ -953,6 +981,7 @@ export async function publishLetters(
   templateName: string,
   lang: Language,
 ) {
+  const configuration = await getConfiguration();
   const urlWithQuery = new URL(
     configuration.julkaiseTuloskirjeetUrl({ hakuOid }),
   );
@@ -967,6 +996,7 @@ export async function sendLetters(
   lang: Language,
   letterId: number,
 ) {
+  const configuration = await getConfiguration();
   const reqBody = {
     hakuOid,
     letterId,
